switch to `Usages` index to using symbol ids for callee, callee_name

+++ jcdb-core/src/main/kotlin/org/utbot/jcdb/impl/features/UsagesExtension.kt
@@ -87,7 +87,7 @@ private suspend fun JcClasspath.findMatches(
             JcClassOrInterfaceImpl(
                 this,
                 it.source
-            ).declaredMethods.filterIndexed { index, jcMethod -> it.offsets.contains(index) }
+            ).declaredMethods.filterIndexed { index, jcMethod -> it.offsets.contains(index.toByte()) }
         }
     }
